+++ src/test/java/de/alpharogroup/collections/map/MapFactoryTest.java
@@ -24,9 +24,9 @@
  */
 package de.alpharogroup.collections.map;
 
-import static org.testng.Assert.assertNotNull;
-import static org.testng.Assert.assertTrue;
 import static org.testng.AssertJUnit.assertEquals;
+import static org.testng.AssertJUnit.assertNotNull;
+import static org.testng.AssertJUnit.assertTrue;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.Comparator;