+++ src/test/java/de/alpharogroup/collections/properties/PropertiesExtensionsTest.java
@@ -24,7 +24,7 @@
  */
 package de.alpharogroup.collections.properties;
 
-import static org.testng.Assert.assertEquals;
+import static org.testng.AssertJUnit.assertEquals;
 import static org.testng.AssertJUnit.assertFalse;
 import static org.testng.AssertJUnit.assertNotNull;
 import static org.testng.AssertJUnit.assertTrue;